+++ botx/models/__init__.py
@@ -1,3 +1,10 @@
+from .botx_api import (
+    BotXTokenResponse,
+    ResponseCommand,
+    ResponseFile,
+    ResponseNotification,
+    ResponseResult,
+)
 from .command_handler import CommandCallback, CommandHandler
 from .common import CommandUIElement, MenuCommand, NotificationOpts, SyncID
 from .cts import CTS, BotCredentials, CTSCredentials
@@ -13,12 +20,6 @@ from .events import ChatCreatedData, UserInChatCreated
 from .file import File
 from .mention import Mention, MentionUser
 from .message import Message, MessageCommand, MessageUser, ReplyMessage
-from .response import (
-    ResponseCommand,
-    ResponseFile,
-    ResponseNotification,
-    ResponseResult,
-)
 from .status import Status, StatusResult
 from .ui import BubbleElement, KeyboardElement
 
@@ -40,6 +41,7 @@ __all__ = (
     "ResponseFile",
     "ResponseNotification",
     "ResponseResult",
+    "BotXTokenResponse",
     "Mention",
     "MentionUser",
     "Message",