+++ docs/changelog.md
@@ -6,6 +6,9 @@
 * Added `HandlersCollector.regex_handler` which accepts a string and compiles it into a regular expression pattern.
 * `MessageCommand.data` field will now automatically converted to events data types corresponding to special events, 
 such as creating a new chat with a bot.
+* Added logging via `loguru`.
+* Renamed `chat_id` argument in `send_file` to `sync_id`.
+* `Bot` and` AsyncBot` can now accept both coroutines and normal functions.
 
 ## 0.11.1
 * Fixed raising exception on successful status codes from the BotX API.